finalizado crud e add ng-bootstrap (confirmation-dialog)

+++ src/app/person-list/person-list.component.ts
@@ -2,7 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { Person } from '../person';
-import { PersonService } from '../person.service';
+import { PersonService } from '../services/person/person.service';
+import { ConfirmDialogService } from '../shared/confirm-dialog/confirm-dialog.service';
 
 @Component({
   selector: 'app-person-list',
@@ -14,28 +15,37 @@ export class PersonListComponent implements OnInit {
   persons!: Observable<Person[]>;
 
   constructor(private personService: PersonService,
-              private router: Router) { }
+              private router: Router,
+              private confirmDialogService: ConfirmDialogService) { }
 
   ngOnInit(): void {
+    this.reloadData();
   }
 
   reloadData(){
     this.persons = this.personService.getPersonList();
   }
 
-  deletePerson(id: number){
-    this.personService.deletePerson(id)
-      .subscribe(
-        data => {
-          console.log(data);
-          this.reloadData();
-        },
-        error => console.log((error));
-      );
+  deletePerson(id: number){   
+    this.confirmDialogService.confirm("DELETE PERSON", "Are you sure?", "Delete", "Cancel")
+      .then(
+        () => {
+          this.personService.deletePerson(id)
+            .subscribe(
+              data => {
+                console.log("to aqui");
+                console.log(data);
+                this.reloadData();
+              },
+              error => console.log((error))
+            );
+        }
+      )
+      .catch( );
   }
 
   personDetails(id: number){
-    this.router.navigate(['details', id]);
+    this.router.navigate([`/person/details`, id]);
   }
 
 }
